Use long long instead of int

+++ fesenko.ivan/S5/createTreeWithTraverses.h
@@ -5,16 +5,16 @@
 namespace fesenko
 {
   template< typename F >
-  using cmd_t = F (AVL< int, std::string, std::less< > >::*)(F f);
+  using cmd_t = F (AVL< long long, std::string, std::less< > >::*)(F f);
   template< typename F >
   using trvrs = AVL< std::string, cmd_t< F >, std::less< > >;
   template< typename F >
   decltype(auto) createTreeWithTraverses()
   {
     trvrs< F > result;
-    result.insert({"ascending", &AVL< int, std::string, std::less< > >::traverse_lnr});
-    result.insert({"descending", &AVL< int, std::string, std::less< > >::traverse_rnl});
-    result.insert({"breadth", &AVL< int, std::string, std::less< > >::traverse_breadth});
+    result.insert({"ascending", &AVL< long long, std::string, std::less< > >::traverse_lnr});
+    result.insert({"descending", &AVL< long long, std::string, std::less< > >::traverse_rnl});
+    result.insert({"breadth", &AVL< long long, std::string, std::less< > >::traverse_breadth});
     return result;
   }
 }
